requirements.txt
fastapi==0.115.12
fastmcp==2.3.3
uvicorn[standard]==0.30.0
motor==3.4.0
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0
pydantic==2.7.1
